Loading the PT_TLS section
application ELF contains the TLS section as well. Since kernel ELF doesnt have it, it is not loaded when the kernel is loaded into memory at boot time. Need to add this case to the switch statement so that it falls through to the next case which is the PT_LOAD. This way PT_TLS will also be loaded