Add in_user per thread variable with getter and setters
adding a per thread variable which will allow us to tell if a UKL thread is currently executing user code or kernel code. This is necessary because on interrupts and faults etc, the CS value pushed on stack will always be kernel CS value even if user code was being executed. This is because we are always in kernel mode in UKL. But we need to take proper actions based on which code we came from i.e., on coming from user code we need to take care of RCU and on way back to user code we need to check if scheduling needs to happen etc etc. But we dont have to do these things on going back to kernel code. So this per thread variable allows us to see, in case of exceptions, which code we came from and take proper action. The functions are just getters and setters for this variable. It takes the value 0 if it is non ukl thread, 1 if UKL thread coming from kernel code and 2 if it is UKL thread coming from user code.