entry_64.S changes: assembly code for kernel entry/exit
This is the low level assembly code which does entry/exit. In each entry point, we check what is the value of the per thread variable called in_user. If its 0, we dont do anything else and let the normal kernel code handle it because its a non UKL thread. If it is 2, we know its a UKL thread which came from user code. Also the CS value pushed on stack will always be kernel CS because UKL is always in kernel mode. The rest of the code makes decisions based on whats the CS value on stack, i.e., deal as a kernel thread or user thread. Instead of reimplementing that logic, we simply change the CS value on stack to be the user CS value, set the in_user variable to be 1 (because we are in kernel code now) and move on. If the in_user value is 1, we dont need to do anything because the CS value on stack is already correct.

On the way back from interrupts or exceptions, we do the opposite. If the saved CS value on stack is user CS, we know we are going back to user code. If we go ahead, the iret will take us to user mode which we dont want. So we update the CS value on stack to be kernel CS and also change the in_user variable to 2 to record that we are goint to user code.

When changing the CS value on stack to user CS and back, we dont actually change it to USER_CS which is 0x33. We change it to a new, somewhat arbitrary value of 0xC3. This has a couple of advantages, 1) most of the entry/exit code written in assembly check the CS value against 0x3. Both 0x33 and 0xC3 pass those tests so once we change the CS value on stack, the rest of the code goes unmodified. The second advantage is that if we ever want to differentiate between user thread and ukl user thread, we can shift right 4 bits and do an AND with 0x3 and get the answer. As the code has matured, we have stopped doing the shift and AND, and simple test against 0x33 and get the answer. So this value is somewhat arbitrary now, any value ending in 0x3 which is not 0x33 should work. We havent changed it, because it also just works.

In UKL, if we are out of user stack, we get a page fault but since we are in kernel mode, no stack switch to kernel stack happens. The hardware tries to push the iret frame on the stack, but since there is already no stack, we get a double fault. Double fault has a stack from IST so we are guaranteed to have a good stack. So in the double fault handler, we put an error value on stack and jump to page fault handler. This way we handle the page fault and iret back to the user code and user stack. We have hijacked the double fault handler, but the idea is that if you get a double fault, you are not going to come back anyway and are going to panic. Later we add a patch which handles page faults through IST stacks, i,e, whenever there is a page fault, you are guaranteed to go to a new stack but that is a system wide change so all page faults are handled that way and might be a performance issue. So we have both options, which one we cant to chose based on config options

This code also has the ukl_entry_syscall_64 entry point, which is called by all the UKL application code whenever it needs to do what used to be system calls.

adding UKL.a to the script which links the final vmlinux binary. It will give an error if it cant find the UKL.a so need to fix that at some point. UKL.a is the partially linked application binary which also contains glibc and other libraries