New variables from Linker script
defining new symbols __tls_data_start, __tls_data_end, __tls_bss_start, __tls_bss_end. These are defined in the linker script and by defning them as externs here, we can use them in the rest of the code. These symbols are used in the initialization to mark read only segments as read only and others as non executable etc. All of this is already done in kernel code, but because we added new sections to the linker script, we need to update the boundaries of read only and non exec areas and these symbols are used there instead of the old ones.