Update user_mode function
We are modifying the user_mode function as well which is also used to check permissions. It is used in the page fault code to check if the faulting thread came from kernel or user mode by checking the CS value in the pt_regs struct. If it is 0x10 i.e. KERNEL_CS, thats a problem because the code does not expect a thread running in kernel mode to get a page fault. But, UKL is always in kernel mode no matter if we are running user code or kernel code. If we dont modify the user_mode function, no page fault from UKL threads will be serviced. So in user_mode function, we check if the cs value in pt_regs is 0xC3 i.e. the value we wrote there when the UKL thread faulted while running user code. If it is )xC3, we simply return 1, which means we came from user code.