Update read only and non-executable marking code
This is init code which marks some parts of the kernel binary read only and some parts no executable. Since we changed the kernel linker script and added user specific sections, these boundaries need to change. So we use new variables from the linker script to update these boundaries. Need to fix the boundaries for freeing kernel pages. Haven't done that yet, so pages are not freed in UKL mode from kernel image