Signal handling updates
Signal handling in UKL is almost the same as normal Linux, except that we need to play with the CS values on stack a little. When a new signal frame is being contructed to iret/sysret to the signal handler, the CS value written in the new frame is USER_CS. This makes sense for normal linux because it knows its going to the signal handler which is in user mode. In UKL, we never leave the kernel mode so instead of putting USER_CS there, we put 0xC3 which is a UKL specific user cs value. While resoring the previous context, we again have to make sure that we dont put USER_CS values on the stack.

Also, while contructing the signal frame on stack, the previous frame is left as is and the new one is created below it. This is because we can come back to the original frame once we return back from the signal handling. Care needs to be taken not to over write the previous frame or part of it with the new frame. In UKL, you get syscall functionality by calling the entry point and you actually return from it instead of sysret or iret. In order to do this, we need to push eflags on the stack and fixup the return address on stack, so our frame is actually a little longer than normal frame. So we change the code a little that if it is a ukl thread, leave a little more space for the previous frame before creating the new frame.