Init changes
As the kernel boots up, it wants to create the first userspace process or pid 1. In UKL mode, this means creating theprocess that will run in kernel mode and run the UKL workload. On top of that, we also want to run a side car which is a full blown Linux userspace so that we can run userspace initialization scripts and do perf or other linuxy stuff. So thats why, instead of creating just the pid 1, we create two processes, ukl and normal userspace.

We copy the command line arguments meant for pid 1 so both processes can get it, and we create one more kernel thread which later exec to create the normal userspace. The only difference between the thread we are on, i.e., the one which later execs to run UKL workload and the new thread which execs userspace is a small one in terms of implementation. Before the UKL thread execs to create and run UKL, we simply change the value of the in_user variable from 0 to 1 i.e., change it from a non ukl thread to a ukl thread running kernel code. If we dont do this, the exec paths for both will be exactly similar and we wont be able to create a ukl thread which stays in kernel mode. More on exec paths in appropriate commits.