Stopping pid 1 exit
This code is more of a convenience that any real functional value. Since the UKL thread is PID 1, if it exits, the kernel panics. But it is normal for user code to exit once it finishes running. And since we use unmodified user code in UKL, the pid 1 can realistically try to exit result in panic. So in the exit code we check if pid 1 is trying to exit. Note that we do this at a point just before finally exiting, and all resources etc have already been freed. So at the very end, we catch it in a while loop, mark it interruptible and call schedule so that we dont panic and other proceses can keep on running. It is especially useful if the side car is running as well and once the UKL workload finishes, you need to export the results out of the machine or ssh into the machine. So all of that continues to run because the PID 1 hasnt exited and kernel hasnt paniced.