exec mechanism in UKL
In UKL, we try to follow the normal exec path as closely as possible. This is because applications expect the process to be set up properly in order to run stably. There are only two difference in normal exec and UKL exec. One, UKL exec does not need to load any elf into memory because its code and data and other segments are already in memory along with the rest of the kernel image. Second, we dont want to load user segment register values i.e., cs, ds, fs, gs etc. If we do that, we will end up ireting to user mode, we want to stay in kernel mode. To this end, all the changes in the exec path are basically if conditions around code which handles elfs and changes segment registers.

If we are a ukl thread, i.e., in_user is not zero, we dont load the eld in memory because that particaluar code is in an if condition. We dont do elf parsing etc. Note that we dont write any new code, just stop the UKL thread from executing parts of exec code which deal with the elf. Since there is no elf, we dont know where to start the heap from because it normally starts after the end of elf binary in memory. So we choose an arbitrarily low value, which just works. Also, since there is no elf, we cant find an entry point to iret into after exec, we hardcode it to jump to _start which we know is the glibc entry point.

When it comes to creating a new mm struct which essentially means creating a new address space, we skip that and intead use init_mm. init_mm is the mm struct which only has the kernel pages. By using that, we make sure that we can access any kernel page without having to first change the cr3 register i.e., without having to load new page tables. Also, if a kthread needs to access memory that is supposed to be on the kernel stack of a ukl thread, it can. This makes more sense in the later optimized versions of UKL where stack is not switched from user to kernel during syscalls. So any reference that was supposed to be on ukl kernel stack is now on ukl user stack and if some other kthread needs to access that reference (e.g., in response to an interrupt etc.) it can still access it. If we create a new mm struct, we will get page not present panics.

Finally, as mentioned above, the code which updates the segment registers with user values is not run because there are condition checking around it. This allows UKl threads to keep the kernel segment register values because UKL has to stay in kernel mode