TLB flush code with init_mm
This is a small condition which is checked in tlb flush code. The conditon makes sure that if you are a user thread, your mm should not be init_mm. We update the condition to make sure that it is only checked for non ukl threads because ukl threads can be running user code and still have init_mm as the mm.