Signal handling in CONFIG_UKL_SAME_STACK
Signal handling in CONFIG_UKL_SAME_STACK is very straightforward. Since we are always on a user stack no matter if we are running user code or kernel code, we dont have to create new stack frames for signal handling. Stack frames help us iret/sysret back from kernel into user code by changing the stack to user stack in the process. Since we are always on user stack, we can avoid this hassle. Whenever a signal is pending, we simply create a function pointer to the signal handler and call it like a normal function. The return address goes on the stack and once the signal handling is done, a simple return pops the return address off of the stack and you are again in a state as if signal handling never happened. Now if you return again, you go back to the original user code which made the syscall that landed you into kernel code.