SAME_STACK page fault deadlock issue
There is a way UKL_SAME_STACK can get stuck in a deadlock. This code fixes that. Imagine if a UKL thread is changing memory i.e., in an mmap call or some related functionality. It has to take a write lock of the mm_struct while it modifies the list of VMAs etc. In normal linux, a thread would only be dealing with VMAs and mmaping things etc while in kernel code on a kernel stack, so there is no way it can suffer a page fault at that point. Imagine a UKL thread on user stack is going through VMAs or mmap functionality etc., it has taken a write lock of mm_struct. During that, it runs out of stack and suffers a page fault. It will immediately go to the page fault handler, which will try to see if the faulting address belongs to a VMA so that it can service it. In order to read the list of VMAs and compare the faulting address against, it needs to take the read lock on the mm_struct. But it cant do that because someone already has a write lock on it, which actually is this very same thread. So it will wait to get the read lock, mark itself interruptible and schedule out, hence the deadlock.

The idea to fix this was given by Larry Woodman. This deadlock will only happen if the thread runs out of stack, there is no other case in UKL that a thread suffers a page fault while in kernel code. So if we save the VMA a reference to the VMA of the stack somewhere, we dont actually have to go through the list of VMAs to find out which VMA is the stack VMA. So whenever there is a pagefault, the first thing we check is if the faulting address belongs to the stack vma through the saved reference to it. If yes, we dont need to take the read lock on mm_struct, we simply read go ahead and service the page fault. If it does not belong to the stack VMA, we know there cant be a deadlock so we go ahead and normally traverse the list of VMAs after taking a read lock on mm_struct first.

This patch deals with that code. We create a per thread variable called user_stack_vma, this will save the reference to the stack vma of each thread. When a thread is created, before it returns to user code for the first time, we call a function which finds out its stack vma and saves it in the user_stack_vma variable. This means every thread, UKL or not, will have this variable correctly populated as soon as they are created. Non ukl threads never use this variable, so it doesnt modify their behavious. They go through the fault handler similar to normal linux. The ukl threads, when they suffer a page fault, use this variable to avoid a deadlock.