UKL_SAME_STACK clone changes
When creating ukl threads through clone in the UKL_SAME_STACK mode, we face a problem. The clone code needs to know where the parent thread stored its registers etc when the clone call was made. This register state needs to be copied as is to the child thread which is being created. Normally, this state is stored on the kernel stack of the parent thread when any syscall is made. This state is restored back before the thread returns to user code.

In UKL_SAME_STACK, ukl threads never change stacks when going to kernel code. So, whenever what used to be a syscall is called, including a clone call, the parent's state is stored on the user stack. The clone code has no idea of this and has no reference to access the user stack. It copies over garbage from parent's kernel stack and the child thread goes into undefined state, crashing the system.

To fix this, we create a new clone flag called CLONE_UKL. In the fork.c, upper 32 bits of the clone flags are zeroed out because they are never used in normal linux. But CLONE_UKL actually lies in the upper 32 bits because all the values below 32 bits are taken. So we fix that by not zeroing out the upper 32 bits of clone flags,

In the syscall entry point for ukl i.e., ukl_entry_syscall_64, we save the pointer to the correct place where parent creates its pt_regs struct on the user stack. Where do we save it? Normally in Linux, in the syscall entry point, there is an instance where the thread saves its user stack pointer in the TSS temporarily, changes its rsp to kernel stack pointer and then retrieves the user stack pointer from the scratch space in TSS and saves it on the stack along with the rest of the iret frame and pt_regs. In UKL same stack, since we dont change the stack, that scratch space is not used at all. Once we create pt_regs on user stack, we save the pointer to the head of the pt_regs (which is on user stack) to this scratch space.

In the clone code, we simply see if the UKL_CLONE flag is true, i.e., we are going to create a UKL thread, not a normal thread. So we simply read off the address of parent's pt_regs from the space in TSS and use it to create the pt_regs for the child.