UKL bypass functions
Whatever UKL version we are running, same stack or otherwise, we can always call the internal syscall-function directly without going through the kernel entry/exit code. This is an optimization because kernel entry/exit code does a lot of housekeeping and other tasks i.e., scheduling or RCU or signal handling etc which takes time. Not doing that makes the syscall-functionality faster.

So we need a way to call the internal functions e.g., instead of a read syscall-function, we directly call the underlying ksys_read. So, we change the syscall wrappers and add stubs into the macros. This way we dont need to know the name of the underlying functions. We simply call ukl_read instead of read. ukl_read is actually a stub which calls the real function. Since we modify the SYSCALL_DEFINE macros, these stubs are automatically created wherever syscalls are defined in the linux code.

But this means that we will never change stack when calling these stubs directly, because stack change only happened in entry/exit code. In non UKL_SAME_STACK case, this is undesirable. In that case, we always change to kernel stack when running kernel code. So we also provide two functions in the entry_64.S file. One of these functions changes the stack from user to kernel, and the other does the reverse procedure. User to kernel switch function is added to the syscall macros just before the underlying function is calles, and kernel to user swtich function is called right after the internal kernel function returns. Great care is taken that these two functions, written in assembly, dont change any registers or any other state except rsp registers. This way, stack changes happen transparently and the code runs smoothly. Any change of state would possbly mean that the kernel returns to the application code in a state which is not consistent with how it was left. In syscalls, apps expect no change of state except that the rax register has the return value.

Adding support functions for UKL Bypass option

These functions are called from glibc. ukl_set_bypass_limit is called to set the threshold for the number of syscalls bypassed before one goes through the normal route. This number can be low, i.e., less benefit but rcu and other book keeping can be done frequently. This number can be high i.e., greater benefit but book keeping will be done less frequently. Because of that, tail latency might go up because some syscalls will take a lot longer because a lot of backlog of book keeping needs to be done. Also, after some value, the increase in performance improvement might not be noticeable so no need to have this number very high.

ukl_set_bypass_syscall sets a per thread variable so syscalls start getting bypassed. After a set number of syscalls have been bypassed, one goes through normally.

ukl_get_bypass_syscall is called in glibc to check if syscall needs to be bypassed. This is where the counter is also incremented. Once the limit is reached, this returns zero so one goes through normally and the counter is reset.