UKL_USE_RET code
This is a further optimization to UKL. We know irets are more expensive than normal rets. Syscalls already use ret but faults and interrupts still use iret. It makes sense to do that if you are returning to user mode, but UKL threads never leave kernel mode. It also makes sense to do it to return to kernel mode, because irets atomically change stack, update rip and update flags. We do all of this and try to do it close to atomically.

This is an optional optimization, one can not turn it on if the slight speed up is not needed. This is only done when UKL threads are returning to user mode. It does not apply to any threads returning to kernel mode or non ukl threads returning to user mode. They still follow the normal path.

Before a UKL thread returns to user_mode, we push the return address and eflags on to the user stack. Then we change the rsp to user stack, pop the flags and do a simple return.