UKL_CREATE_AFTERSPACE
This patch is totally optional, only provides convenience in running UKL. Sometimes while running UKL, we want to kill the sidecar process so that UKL workload runs in a low noise environment. Once UKL workload finishes, we want the sidecar to start again so that we can ssh into the system and export results etc. This code does exactly that, creates a new sidecar i.e., userspace process once UKL exits. Right now it runs init again, but any new script can be run as well.

We create a new kthread at the boot stage and make it wait behind a completion. This completion depends on PID 1 or UKL exiting. Before UKL goes into a while loop and calls the scheduler in the exit code, it marks the completion as complete. This triggers the waiting kthread to exec init and create the sidecar again.